requirements.txt
requests
pandas
python-dotenv
psycopg2-binary
click
pytest
tenacity
sqlalchemy